More tests running on netfx
